+++ CHANGELOG.rst
@@ -1,4 +1,4 @@
-HEAD (2023-02-??)
+0.22 (2023-03-13)
 -----------------
 
 *	Added incremental loading of layout controls.